+++ src/util/test/util.ts
@@ -252,7 +252,6 @@ export function createTerrain(): Terrain {
         getMinTileElevationForLngLatZoom: () => 0,
         resetElevationCache: () => {},
         getFramebuffer: () => ({}),
-        depthAtPoint: () => .9,
         tileManager: {
             update: () => {},
             getRenderableTiles: () => [],